+++ miio/gateway/__init__.py
@@ -2,3 +2,7 @@
 
 # flake8: noqa
 from .gateway import Gateway
+from .alarm import GatewayAlarm
+from .light import GatewayLight
+from .radio import GatewayRadio
+from .zigbee import GatewayZigbee
