+++ miio/gateway/devices/light.py
@@ -4,7 +4,7 @@ import attr
 import click
 
 from ...click_common import command
-from .sub_device import SubDevice
+from .subdevice import SubDevice
 
 
 class LightBulb(SubDevice):
