--- miio/gateway/gateway_device/__init__.py
@@ -1,7 +1,0 @@
-"""Xiaomi Gateway device base class."""
-
-# flake8: noqa
-from .gateway_alarm import GatewayAlarm
-from .gateway_light import GatewayLight
-from .gateway_radio import GatewayRadio
-from .gateway_zigbee import GatewayZigbee
