+++ miio/gateway/light.py
@@ -4,9 +4,9 @@ from typing import Tuple
 
 import click
 
-from ...click_common import command
-from ...utils import brightness_and_color_to_int, int_to_brightness, int_to_rgb
-from .gateway_device import GatewayDevice
+from ..click_common import command
+from ..utils import brightness_and_color_to_int, int_to_brightness, int_to_rgb
+from .gatewaydevice import GatewayDevice
 
 color_map = {
     "red": (255, 0, 0),
